Moved links generation to paging-behavior object
