+++ crnk-core/src/main/java/io/crnk/core/queryspec/pagingspec/OffsetLimitPagingBehavior.java
@@ -1,7 +1,12 @@
 package io.crnk.core.queryspec.pagingspec;
 
+import io.crnk.core.engine.query.QueryAdapter;
 import io.crnk.core.exception.BadRequestException;
 import io.crnk.core.exception.ParametersDeserializationException;
+import io.crnk.core.resource.links.PagedLinksInformation;
+import io.crnk.core.resource.list.ResourceList;
+import io.crnk.core.resource.meta.HasMoreResourcesMetaInformation;
+import io.crnk.core.resource.meta.PagedMetaInformation;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -60,6 +65,88 @@ public class OffsetLimitPagingBehavior implements PagingBehavior<OffsetLimitPagi
 	@Override
 	public OffsetLimitPagingSpec createEmptyPagingSpec() {
 		return new OffsetLimitPagingSpec(defaultOffset, defaultLimit);
+	}
+
+	@Override
+	public void build(final PagedLinksInformation linksInformation,
+					  final ResourceList<?> resources,
+					  final QueryAdapter queryAdapter,
+					  final PagingSpecUrlBuilder urlBuilder) {
+		Long totalCount = getTotalCount(resources);
+		Boolean isNextPageAvailable = isNextPageAvailable(resources);
+		if ((totalCount != null || isNextPageAvailable != null) && !hasPageLinks(linksInformation)) {
+			// only enrich if not already set
+			boolean hasResults = resources.iterator().hasNext();
+			doEnrichPageLinksInformation(linksInformation, totalCount, isNextPageAvailable,
+					queryAdapter, hasResults, urlBuilder);
+		}
+	}
+
+	@Override
+	public boolean isRequired(final OffsetLimitPagingSpec pagingSpec) {
+		return pagingSpec.getOffset() != 0 || pagingSpec.getLimit() != null;
+	}
+
+	private void doEnrichPageLinksInformation(PagedLinksInformation linksInformation, Long total,
+											  Boolean isNextPageAvailable, QueryAdapter queryAdapter,
+											  boolean hasResults,
+											  PagingSpecUrlBuilder urlBuilder) {
+		OffsetLimitPagingSpec offsetLimitPagingSpec = (OffsetLimitPagingSpec) queryAdapter.getPagingSpec();
+		long pageSize = offsetLimitPagingSpec.getLimit();
+		long offset = offsetLimitPagingSpec.getOffset();
+		long currentPage = offset / pageSize;
+		if (currentPage * pageSize != offset) {
+			throw new BadRequestException("offset " + offset + " is not a multiple of limit " + pageSize);
+		}
+		if (total != null) {
+			isNextPageAvailable = offset + pageSize < total;
+		}
+
+		if (offset > 0 || hasResults) {
+			Long totalPages = total != null ? (total + pageSize - 1) / pageSize : null;
+			QueryAdapter pageSpec = queryAdapter.duplicate();
+			pageSpec.setPagingSpec(new OffsetLimitPagingSpec(0L, pageSize));
+			linksInformation.setFirst(urlBuilder.build(pageSpec));
+
+			if (totalPages != null && totalPages > 0) {
+				pageSpec.setPagingSpec(new OffsetLimitPagingSpec((totalPages - 1) * pageSize, pageSize));
+				linksInformation.setLast(urlBuilder.build(pageSpec));
+			}
+
+			if (currentPage > 0) {
+				pageSpec.setPagingSpec(new OffsetLimitPagingSpec((currentPage - 1) * pageSize, pageSize));
+				linksInformation.setPrev(urlBuilder.build(pageSpec));
+			}
+
+
+			if (isNextPageAvailable) {
+				pageSpec.setPagingSpec(new OffsetLimitPagingSpec((currentPage + 1) * pageSize, pageSize));
+				linksInformation.setNext(urlBuilder.build(pageSpec));
+			}
+		}
+	}
+
+	private Long getTotalCount(ResourceList<?> resources) {
+		PagedMetaInformation pagedMeta = resources.getMeta(PagedMetaInformation.class);
+		if (pagedMeta != null) {
+			return pagedMeta.getTotalResourceCount();
+		}
+
+		return null;
+	}
+
+	private Boolean isNextPageAvailable(ResourceList<?> resources) {
+		HasMoreResourcesMetaInformation pagedMeta = resources.getMeta(HasMoreResourcesMetaInformation.class);
+		if (pagedMeta != null) {
+			return pagedMeta.getHasMoreResources();
+		}
+
+		return null;
+	}
+
+	private boolean hasPageLinks(PagedLinksInformation pagedLinksInformation) {
+		return pagedLinksInformation.getFirst() != null || pagedLinksInformation.getLast() != null
+				|| pagedLinksInformation.getPrev() != null || pagedLinksInformation.getNext() != null;
 	}
 
 	private Long getValue(final String name, final Set<String> values) {
