+++ crnk-core/src/main/java/io/crnk/core/queryspec/pagingspec/PagingBehavior.java
@@ -1,5 +1,9 @@
 package io.crnk.core.queryspec.pagingspec;
 
+import io.crnk.core.engine.query.QueryAdapter;
+import io.crnk.core.resource.links.PagedLinksInformation;
+import io.crnk.core.resource.list.ResourceList;
+
 import java.util.Map;
 import java.util.Set;
 
@@ -28,4 +32,24 @@ public interface PagingBehavior<T extends PagingSpec> {
 	 * @return New {@link PagingSpec} instance
 	 */
 	T createEmptyPagingSpec();
+
+	/**
+	 * Fills out the paging links
+	 *
+	 * @param linksInformation {@link PagedLinksInformation} instance
+	 * @param resources {@link ResourceList} of resources
+	 * @param queryAdapter {@link QueryAdapter} instance
+	 * @param urlBuilder {@link PagingSpecUrlBuilder} instance to provide a way to build a link
+	 */
+	void build(PagedLinksInformation linksInformation, ResourceList<?> resources,
+			   QueryAdapter queryAdapter,
+			   PagingSpecUrlBuilder urlBuilder);
+
+	/**
+	 * Determines whether Crnk needs to provide paging links via {@link #build(PagedLinksInformation, ResourceList, QueryAdapter, PagingSpecUrlBuilder)}
+	 *
+	 * @param pagingSpec {@link PagingSpec} instance
+	 * @return True in case of pagination is required otherwise False
+	 */
+	boolean isRequired(T pagingSpec);
 }
